[monodroid] Allow sideloading real paths in addition to in-memory assemblies
This commit extends the previous work from https://github.com/xamarin/xamarin-android/pull/3365 that allowed in-memory data to be given for an assembly content to also permit supplying a real path instead.

The reasoning is that for some assemblies it's safe and preferable to allow them to be loaded from a given location. A good example are assemblies coming from the NuGet global package cache which are considered read-only and thus won't cause Windows file-locking issues when loaded up (as opposed to when they are loading from the project obj directory).

By selectively allowing those type of assemblies to be loaded from disk, we avoid putting unnecessary memory pressure on the system while retaining the flexibility to do so if we choose.